libext: add function extract_time libdb: import time to db

+++ src/lib/ext/ext.h
@@ -1,11 +1,11 @@
-/** @ext-instr.h Contains functions used to extract instruments from mid files */
+/** @ext.h Contains functions used to extract data from mid files */
 
 #ifndef _EXT_H_
 #define _EXT_H_
 
 #include <mid/mid.h>
 
-int extract_instr_classes(mid_t *mid);
-int extract_time(mid_t *mid);
+unsigned extract_instr_classes(mid_t *mid);
+double extract_time(mid_t *mid);
 
 #endif
